info in desc.
make a start position for each scene and load that based on scene index

put in a new text field called PopUp and create public method in PlayerManager for setting the PopUp text

also change the font to something more bearable to look at for all the text (except dev buttons)

Disable movement controls [Done]
    but a bool into the movement manager or whatever controls the controls
    link to player through player. Movement Manager

also create a command in the console to disable or enable movement commands. I figure the players should be able to run all the commands even if they're not necessarily useful.